+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Cacheing of F-Droid package index to local config directory
+
 ## [0.7.0] - 2021-11-29
 ### Added
 - Ability to download versioned apps on APKPure and F-Droid, as well as making it possible in a separate field if you're using a CSV